+++ poseidon/src/lib.rs
@@ -1,12 +1,11 @@
 use ark_bls12_381::Fr;
 use ark_relations::r1cs::{ConstraintSystemRef, SynthesisError};
-use ark_sponge::{
-    constraints::CryptographicSpongeVar,
-    poseidon::{constraints::PoseidonSpongeVar, PoseidonParameters as ArkPoseidonParameters},
-};
+use ark_sponge::{constraints::CryptographicSpongeVar, poseidon::constraints::PoseidonSpongeVar};
 use once_cell::sync::Lazy;
-use poseidon_paramgen::{Alpha, PoseidonParameters};
+use poseidon_paramgen::PoseidonParameters;
+use utils::to_ark_sponge_parameters;
 
+mod utils;
 pub mod parameters {
     include!(concat!(env!("OUT_DIR"), "/parameters.rs"));
 }
@@ -14,29 +13,9 @@ pub mod parameters {
 pub type CircuitField = ark_bls12_381::Fr;
 pub type FpVar = ark_r1cs_std::fields::fp::FpVar<CircuitField>;
 
-/// Parameters for the rate-1 instance of Poseidon
+/// Parameters for the 1 to 1 instance of Poseidon
+// TODO : use lazy to assign ark parameters
 pub static RATE_1_PARAMETERS: Lazy<PoseidonParameters<Fr>> = Lazy::new(parameters::rate_1);
-
-fn convert_to_ark_sponge_parameters(params: PoseidonParameters<Fr>) -> ArkPoseidonParameters<Fr> {
-    let alpha = match params.alpha {
-        Alpha::Exponent(exp) => exp as u64,
-        Alpha::Inverse => panic!("ark-sponge does not allow inverse alpha"),
-    };
-    let capacity = 1;
-    let rate = params.t - capacity;
-    let full_rounds = params.rounds.full();
-    let partial_rounds = params.rounds.partial();
-
-    ArkPoseidonParameters {
-        full_rounds,
-        partial_rounds,
-        alpha,
-        ark: params.arc.into(),
-        mds: params.mds.into(),
-        rate,
-        capacity,
-    }
-}
 
 /// hashes one field value inside the circuit
 pub fn one_to_one_hash(
@@ -44,7 +23,8 @@ pub fn one_to_one_hash(
     domain_separator: &FpVar,
     value: FpVar,
 ) -> Result<FpVar, SynthesisError> {
-    let ark_parameters = convert_to_ark_sponge_parameters(RATE_1_PARAMETERS.clone());
+    // TODO : omit this lazily instantiate sponge params instead
+    let ark_parameters = to_ark_sponge_parameters(RATE_1_PARAMETERS.clone());
     let mut state: PoseidonSpongeVar<Fr> = PoseidonSpongeVar::new(cs, &ark_parameters);
 
     state.absorb(&vec![domain_separator, &value])?;
