+++ poseidon/src/utils.rs
@@ -1,0 +1,27 @@
+use ark_bls12_381::Fr;
+use ark_sponge::poseidon::PoseidonParameters as ArkPoseidonParameters;
+use poseidon_paramgen::{Alpha, PoseidonParameters};
+
+// taken from Penumbra (https://github.com/penumbra-zone/poseidon377/blob/a2d8c7a3288e2e877ac88a4d8fd3cc4ff2b52c04/poseidon377/src/r1cs.rs#L12)
+pub(crate) fn to_ark_sponge_parameters(
+    params: PoseidonParameters<Fr>,
+) -> ArkPoseidonParameters<Fr> {
+    let alpha = match params.alpha {
+        Alpha::Exponent(exp) => exp as u64,
+        Alpha::Inverse => panic!("ark-sponge does not allow inverse alpha"),
+    };
+    let capacity = 1;
+    let rate = params.t - capacity;
+    let full_rounds = params.rounds.full();
+    let partial_rounds = params.rounds.partial();
+
+    ArkPoseidonParameters {
+        full_rounds,
+        partial_rounds,
+        alpha,
+        ark: params.arc.into(),
+        mds: params.mds.into(),
+        rate,
+        capacity,
+    }
+}
